## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

This PR https://github.com/StarRocks/starrocks/pull/4934 causes the performance regression on branch-2.2

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

vector::assign(n, 1) is needless and time-consuming so remove it, and make the for-loop of bloom-filter probing more tight.
the performance test as follows.
```
+===================================+========+===================+================+
| query                             | 2.2.1  | 2.2.2(regression) | this PR(2.2.2) |
+===================================+========+===================+================+
| join.18_ssb_inner.q4.q_18_4_4.sql | 0.224s | 0.350s            | 0.223s         |
| join.19_ssb_left.q4.q_19_4_4.sql  | 0.234s | 0.336s            | 0.219s         |
+-----------------------------------+--------+-------------------+----------------+
```