@satanson 2022-06-24
run starrocks_fe_unittest