@wanpengfei-git 2022-06-24
run starrocks_admit_test